move elevenlabs to speech service, more work on twitter stuff
